desloppify: move xmlrpc value to leaf module
