+++ src/xmlrpc/call.rs
@@ -2,7 +2,7 @@
 
 use std::collections::BTreeMap;
 
-use crate::xmlrpc::Value;
+use crate::xmlrpc::value::Value;
 
 /// Escape XML special characters in text content.
 fn xml_escape(s: &str) -> String {
